Move all editors to the editors folder.
