+++ src/content/plugins/crm.js
@@ -2,7 +2,7 @@
  */
 
 import {parseTemplate} from '../utils.js';
-import {insertTemplate} from '../utils/editor-generic.js';
+import {insertTemplate} from '../editors/editor-generic.js';
 
 function getVariableAttribute (element) {
     const attributes = [ 'data-gorgias-variables', 'data-briskine-variables' ];
